+++ packages/types/custom/core.ts
@@ -649,6 +649,7 @@ export interface CoreState {
   hovered: number
   language: string
   map: number
+  mapHasDimensions: boolean
   moveHandle: number
   moveHandleActionButton: number
   plugin: object
@@ -656,23 +657,30 @@ export interface CoreState {
   zoomLevel: number
 }
 
-export interface CoreGetters {
+export interface CoreGetters
+  extends Omit<
+    CoreState,
+    | 'components'
+    | 'hovered'
+    | 'map'
+    | 'moveHandle'
+    | 'moveHandleActionButton'
+    | 'selected'
+  > {
+  // omitted from CoreState as actual getter type diverges
   components: PluginContainer[]
-  configuration: MapConfig
+  hovered: Feature | null
+  map: Map
+  mapHasDimensions: boolean
+  moveHandle: MoveHandleProperties
+  moveHandleActionButton: MoveHandleActionButton
+  selected: Feature | null
+
+  // regular getters
   hasSmallHeight: boolean
   hasSmallWidth: boolean
   /** Whether the application currently has the same size as the visual viewport of the users browser */
   hasWindowSize: boolean
-  hovered: Feature | null
-  errors: PolarError[]
-  map: Map
-  moveHandle: MoveHandleProperties
-  moveHandleActionButton: MoveHandleActionButton
-  selected: Feature | null
-  clientHeight: number
-  clientWidth: number
-  center: number[] | null
-  zoomLevel: number
 }
 
 export type PolarGetter<S, G, P> = (