+++ packages/types/custom/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unpublished
 
 - Feature: Add `afterResultComponent` to `AddressSearchConfiguration` for custom search result suffixes.
+- Feature: Add `mapHasDimensions` to `CoreState` and `CoreGetters`.
 - Fix: Add `string` as option for `SearchType` since arbitrary strings can be registered.
 
 ## 1.4.1